Add possibility of timestamped pins messages
This commit adds the possibility to request timestamps with all pins
messages sent from the Poark server.  The timestamp is a 31 bit integer
number (see description below) indicating how many milliseconds have passed
since the server was started.  This gives a wrap around time just short of
600 minutes.

To minimize the overhead of the timestamps, especially when they are not
used, they are transmitted as faked pin readings.  All pins with the top
bit set is considered to be a timestamp and the lower 15 bits to be
combined with the value int to obtain a 31 bit timestamp.

BUG=20
TEST=poark_client should run as usual with the addition of timestamps
being displayed before each set of pin states.