Make the configure message show correct LCD status
The current define used for the LCD display (LCD_DEBUG) is inconsistent
with the other define tunable features (WITH_XXX).  That would be a reason
for changing LCD_DEBUG into WITH_LCD here, but at the same time activating
the LCD also activates a whole lot of debug messages.  This motivates
keeping the inconsistent naming.

BUG=24
TEST=Request a configure report with LCD activated and make sure it is
correctly reported as turned on.