Refactor LCD debug output
This commit refactors some quite frequently used LCD debug output
code.  I generally prefer not to have macros (for once they lack type
checks), but I don't know of a better and as efficient way to do this.

The code relies on macro ellipses (...), which is part of the C99.
This could in some rare cases become a compatibility problem, but I
think the simplicity of the macro definition makes it worth it.

BUG=28
TEST=Previous code should work as before