The length of std_msgs arrays is given in number of array elements
This caused two potentially fatal interacting and difficult to find
errors.  First, the array was assumed larger than it actually was.
Second, the unsigned char used for the length did overflow.  For a
board with 70 pins, enough space for 140 pins was indicated but due to
the overflow the result was 12 pins.

BUG=35
TEST=compilation should pass and Poark should still function.