Use GLCD v3.0 instead of ks0108
The LCD controlling library that is currently used in Poark
  http://www.arduino.cc/playground/Code/GLCDks0108
is now in the process of being superseded by GLCD v 3.0
  http://arduino.cc/forum/index.php/topic,56705.0.html

As we are currently patch the ks0108 library to work correctly with the
Arduino Mega2560, and to kill some warnings.  The GLCD v 3.0 library
should supports the Mega2560 out of the box, hence it is a good idea to
integrate the newer library rather than keep patching the older library.

BUG=37
TEST=The LCD library should work out of the box and all test work as before.