Update poark_client to reflect that MsTimer2 is used
MsTimer2 uses the Mega boards timer2 circuit, changing its resolution to
1 ms.  This changes the PWM cycle frequency, which by default is 31250 Hz,
on pins 9 and 10 which are timer2 driven. The duty cycle should still be
correct but the whole cycle much slower than expected.

To avoid setting a bad example, pick new PWM pins for use in the example.
Pin 2, 3, and 4 control RGB LED3 on the Android DemoKit making them
a good alternative.

BUG=41
TEST=The all 3 LEDs controlled by poark_client should now behave the same.