Adding mode to connect to the Arduino's changed pin interrupts
With this commit it is possible to set pins on the Arduino board that can
fire interrupts into INTERRUPT mode.  In INTERRUPT mode the pin reports how
many times the interrupt has fired during the past sampling period.  No
message is sent unless the interrupt has been activated to minimize the
communication.

The new mode can be used efficiently to monitor a pin for rare changes, or
to make sure no ticks are missed from an angle encoder or similar.

BUG=44
TEST=poark_client should now be able to toggle a LED using a interrupt.